+++ src/core/include/core/GDK/Application.hpp
@@ -4,6 +4,8 @@
 
 namespace gamedevkit {
 
+class AbstractGame;
+class AbstractRenderer;
 class Window;
 
 class Application final {
@@ -19,6 +21,8 @@ public:
 
 public:
     auto window(std::unique_ptr<Window> window) -> Application&;
+    auto game(std::shared_ptr<AbstractGame> game) -> Application&;
+    auto renderer(std::unique_ptr<AbstractRenderer> renderer) -> Application&;
 
 public:
     auto setup() -> void;
@@ -29,6 +33,8 @@ private:
     const std::unique_ptr<Details> details_;
 
     std::unique_ptr<Window> window_;
+    std::shared_ptr<AbstractGame> game_;
+    std::unique_ptr<AbstractRenderer> renderer_;
 };
 
 }  // namespace gamedevkit