+++ tests/test_core_Application.cpp
@@ -1,6 +1,4 @@
-#include <chrono>
-#include <future>
-
+#include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
 #include <glfwcxx/CoreStub.hpp>
@@ -8,8 +6,22 @@
 
 #include <GL/glew.h>
 
+#include <GDK/AbstractGame.hpp>
+#include <GDK/AbstractRenderer.hpp>
 #include <GDK/Application.hpp>
 #include <GDK/Window.hpp>
+
+class MockedGame : public gamedevkit::AbstractGame {
+public:
+    MOCK_METHOD(void, setup, (), (override));
+    MOCK_METHOD(void, update, (), (override));
+};
+
+class MockedRenderer : public gamedevkit::AbstractRenderer {
+public:
+    MOCK_METHOD(void, setup, (std::shared_ptr<gamedevkit::AbstractGame>), (override));
+    MOCK_METHOD(void, render, (), (override));
+};
 
 class gamedevkit_application : public testing::Test {
 public:
@@ -20,10 +32,14 @@ public:
         glfwcxx::WindowStub::reset();
         application_ = std::make_unique<gamedevkit::Application>();
         window_ = std::make_unique<gamedevkit::Window>("", gamedevkit::WindowResolution{100, 100});
+        game_ = std::make_shared<MockedGame>();
+        renderer_ = std::make_unique<MockedRenderer>();
     }
 
     std::unique_ptr<gamedevkit::Application> application_{nullptr};
     std::unique_ptr<gamedevkit::Window> window_{nullptr};
+    std::shared_ptr<MockedGame> game_{nullptr};
+    std::unique_ptr<MockedRenderer> renderer_{nullptr};
 };
 
 TEST_F(gamedevkit_application, throws_runtime_error_on_construction_when_glfwcxx_cannot_be_initialized)
@@ -44,23 +60,28 @@ TEST_F(gamedevkit_application, throws_runtime_error_when_window_was_not_set_befo
     ASSERT_THROW(application_->setup(), std::runtime_error);
 }
 
-TEST_F(gamedevkit_application, throws_runtime_error_when_window_making_context_current_fails)
+TEST_F(gamedevkit_application, throws_runtime_error_when_game_was_not_set_before_setup)
 {
-    glfwcxx::WindowStub::make_context_current_failure();
-    application_->window(std::move(window_));
-    ASSERT_THROW(application_->setup(), std::runtime_error);
+    ASSERT_THROW(application_->window(std::move(window_)).setup(), std::runtime_error);
+}
+
+TEST_F(gamedevkit_application, throws_runtime_error_when_renderer_was_not_set_before_setup)
+{
+    ASSERT_THROW(application_->window(std::move(window_)).game(game_).setup(), std::runtime_error);
 }
 
 TEST_F(gamedevkit_application, throws_runtime_error_when_gl_functions_were_not_initialized)
 {
     gamedevkit::GlewStub::glew_init_return_value(GLEW_OK + 1u);
-    application_->window(std::move(window_));
+    application_->window(std::move(window_)).game(game_).renderer(std::move(renderer_));
     ASSERT_THROW(application_->setup(), std::runtime_error);
 }
 
-TEST_F(gamedevkit_application, successfully_configures_window_when_setup_invoked)
+TEST_F(gamedevkit_application, successfully_configures_window_and_game_and_renderer_when_setup_invoked)
 {
-    ASSERT_NO_THROW(application_->window(std::move(window_)).setup());
+    EXPECT_CALL(*game_, setup).Times(testing::Exactly(1));
+    EXPECT_CALL(*renderer_, setup(testing::_)).Times(testing::Exactly(1));
+    ASSERT_NO_THROW(application_->window(std::move(window_)).game(game_).renderer(std::move(renderer_)).setup());
 }
 
 TEST_F(gamedevkit_application, successfully_runs_two_game_loops_and_returns_exit_success_when_window_should_close)
@@ -73,7 +94,10 @@ TEST_F(gamedevkit_application, successfully_runs_two_game_loops_and_returns_exit
             glfwcxx::WindowStub::close_window();
     });
 
-    application_->window(std::move(window_)).setup();
+    EXPECT_CALL(*game_, update).Times(testing::Exactly(1));
+    EXPECT_CALL(*renderer_, render).Times(testing::Exactly(2));
+
+    application_->window(std::move(window_)).game(game_).renderer(std::move(renderer_)).setup();
 
     // TODO: it might stuck, move it to a separate thread of execution with timeout
     EXPECT_EQ(EXIT_SUCCESS, application_->run());