+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 - Setup development infrastructure [(#1)](https://github.com/EugeneKuznetsov/gamedevkit/issues/1)
 - Setup continuous integration workflow [(#2)](https://github.com/EugeneKuznetsov/gamedevkit/issues/2)
 - Application class for setting up game infrastructure [(#7)](https://github.com/EugeneKuznetsov/gamedevkit/issues/7)
+- Game and Renderer interface [(#9)](https://github.com/EugeneKuznetsov/gamedevkit/issues/9)
 
 ## Release